collect applications on the node
maintaining separate apps when they all depends each others and are part of
the core application is a non-sense. It's exhausting to maintain for each the
right version between devs even with the _checkout folder. Also moving/editing files
between each apps during a long cycle of development is proved to create a lot
of insane situations.

This patch collect for now all "core" applications in apps/ .