fix view changes
if a document has been deleted and removed from the view, we should emit the
changes and mark it as deleted. If the doc has been removed from the changes
mark it as removed.

This changes also introduces the `_removed` member for a revision to distinct
removed documents in a view from deleted documents.