start to remove the ini file
this change for now allows to set the database/view dir via the environment and set the nodeid if it doesn't exist. The
nodeid is stored in the database_dir by default.

Note that the nodeid is the only item that really need to be persisted across restart so it will make easier later the
removal of the ini file.