add barrel_auth actor to keep our secret
The secret used for generating and checking cookies is keep ourside the config in the <datadir/.barrel.secret> file.

This file can only be read by the barrel user. When the file is open, the date and its content are also checked.

While i'm here directly handle hexadecimal in binary instead of doing a roundtrip from binary to hex then hex to
binary...