introduce the new local user handling
Previously barrel was storing the admin accounts in the ini file without much concerns
about the security or persistence issues (during an update the INI could be corrupted.

This changes add a new local user actor to barrel which maintain a db file in the
data directory with the following properties:

- The file is only writable by the db owner
- On update a new file is created.
- If the update fail, the old database directory is reverted

Like the ini file, it gives an api to add/delete/update admins users directly. For now
this api is not exposed in a friendly manner.

Also it ass an API to reload the db file in memory, so anyoen can eventually update it
from the filesystem.