remove the ini configuration (#45)
* start to remove the ini file

this change for now allows to set the database/view dir via the environment and set the nodeid if it doesn't exist. The
nodeid is stored in the database_dir by default.

Note that the nodeid is the only item that really need to be persisted across restart so it will make easier later the
removal of the ini file.

* replace barrel_config in api

* add basic api to start console and api listeners

* our last dep is now in hex.pm

* bump mochicown to 0.6.4

* use the same data folder for views and databases

as of today there is no rela need for having 2 folders

* add barrel_auth actor to keep our secret

The secret used for generating and checking cookies is keep ourside the config in the <datadir/.barrel.secret> file.

This file can only be read by the barrel user. When the file is open, the date and its content are also checked.

While i'm here directly handle hexadecimal in binary instead of doing a roundtrip from binary to hex then hex to
binary...

* refactor supervision with auth

this changes does the following:
- handle the supervision of barrel_auth in barrel_ui_sup the supervisor for any interface to the user
- remove barrel_auth_cache from the lists of the daemon. It's not optionnal at all when you're using the HTTP api.
- move barrel_auth to its own folder

also make sure that the console is started by the api supervisor like it should be.

* introduce the new local user handling

Previously barrel was storing the admin accounts in the ini file without much concerns
about the security or persistence issues (during an update the INI could be corrupted.

This changes add a new local user actor to barrel which maintain a db file in the
data directory with the following properties:

- The file is only writable by the db owner
- On update a new file is created.
- If the update fail, the old database directory is reverted

Like the ini file, it gives an api to add/delete/update admins users directly. For now
this api is not exposed in a friendly manner.

Also it ass an API to reload the db file in memory, so anyoen can eventually update it
from the filesystem.

* fix cookie authentication

* unbreak the console UI

* couch_passwords -> barrel_passwords

* revisit the supervision

This changes introduce the safe supervisor to run servers that don't require a full restart of all the services if they
hangs. Also remove the daemon supervisor.

* unbreak tests

* remove barrel_daemons_sup

* remove most of the barrel_config call

* latest bits for removing the ini from the barrel app

* say good bye to couch_external_manager

* add needed hooks to instrument the compaction daemon.

allows to register/unregister a compaction task dynamically

* add missing copyright header

* list, pause, resume compaction tasks

* rename the compaction daemon to barrel_compaction_daemon

* fix authentication

fix cookie authentication and fully use the concept of resource in cowboy_rest.

* remove anything about ini abd based on the ini configuration

* make sure to not conflict with dabase name using the default config dir

* simpler config for listeners

* hello fast_yaml

* correctly display web uris

forgot to use the new "listen" section

* unbeak tests

* PASSWD to users

* say hello to the new yaml configuration

note: this configuration is completly optionnal and can be omitted by the user if we want to setup a node using a pure
erlang configuration or use its system.

* list cors methods in yaml

* more options