fix api to support multiple stores in one vm
Previously we were expecting to create the db name was unique to a vm and 2 stores created on the same vm should had to use different db names. More over the creation on the second store was silently ignored if the database was already started on the second store. This change fix it by making the connecyion to a database returning an opaque connection object that will be used in place of the db name.

replication, attachments and all call have been adapted. all tests pass.