make sure to send the ancestors in order
spotted by @bernardn revisions weren't sent in order to the target which
sometimes was forbidding a revision to be added to a doc. This change
fix it by sending the revision we found in order.

also fix tests since now we get an {error, not_found} when the doc is
already fixed.

NOTE: a better fix for the tests would be checking first if the doc is
expected to be deleted before really deleting it