need the revision only for concurrency purpose
this patch makes the following change

- "_rev" has not to be part of the doc any more, it can now *optionnaly*
be passed as an option
- "_rev" setting is optionnal and only used for concurrent check. It is
updated on the storage node. In other cases put_rev can be used.
- "post" is now only used for document creation. The "id" property is
required. It is also never dynamically created, so it's really a user
data without specific exception for it.
- add "upsert" option to post to handle "create or update an exiting
docment" actions
- A doc needs to be created by post first, then updated via put. So no
accidental creation can happen
- delete only need the docid now. Like put, the rev is optionnal and
only used for concurrency purpose