simplify the indexing logic
we don't need to know the offset or position, instead we should only
rely on the partial path and the path given during the query to return
the results.

this is a work in progress, for now onl make sure the that we still work
with the current behavior that pass found docs in the function instead
of the fragment.