store meta keys as independant variable
Instead of storing all metadata in one value, we now store them as independant keys. Doing this remove the need to read the info from the db when updating the counter and reduce the size on the disk.

Note: in a next step we may want to use column to store them and use the merge operator to speed inserts.