last write win
By default the last write win now until we pass a revision. We still keep the revision tree for it.

When a doc is updated without passing a revision but the doc is already existing on disk we take the last winning revision as parent . If a revision is given then usual checking is handled.

While I'm here add the is_upsert option to post to allows doc creation to override an exising document. In that case, the last winning revision is used as parent like put above.